+++ code/src/cljs/sixsq/nuvla/ui/credentials/events.cljs
@@ -7,8 +7,8 @@
     [sixsq.nuvla.ui.credentials.spec :as spec]
     [sixsq.nuvla.ui.credentials.utils :as utils]
     [sixsq.nuvla.ui.job.events :as job-events]
-    [sixsq.nuvla.ui.messages.events :as messages-events]
     [sixsq.nuvla.ui.main.spec :as main-spec]
+    [sixsq.nuvla.ui.messages.events :as messages-events]
     [sixsq.nuvla.ui.utils.general :as general-utils]
     [sixsq.nuvla.ui.utils.response :as response]))