+++ code/src/cljs/sixsq/nuvla/ui/edge/events.cljs
@@ -387,3 +387,26 @@
   ::set-active-tab-index
   (fn [db [_ active-tab-index]]
     (assoc db ::spec/active-tab-index active-tab-index)))
+
+
+(reg-event-fx
+  ::enable-host-level-management
+  (fn [_ [_ nuvlabox-id]]
+    {::cimi-api-fx/operation
+     [nuvlabox-id
+      "enable-host-level-management"
+      #(if (instance? js/Error %)
+         (let [{:keys [status message]} (response/parse-ex-info %)]
+           (dispatch [::messages-events/add
+                      {:header  (cond-> (str "error enabling host level management for " nuvlabox-id)
+                                        status (str " (" status ")"))
+                       :content message
+                       :type    :error}]))
+         (dispatch [::set-nuvlabox-playbooks-cronjob %]))
+      nil]}))
+
+
+(reg-event-db
+  ::set-nuvlabox-playbooks-cronjob
+  (fn [db [_ cronjob]]
+    (assoc db ::spec/nuvlabox-playbooks-cronjob cronjob)))