+++ code/src/cljs/sixsq/nuvla/ui/edge/utils.cljs
@@ -1,8 +1,7 @@
 (ns sixsq.nuvla.ui.edge.utils
   (:require
     [clojure.string :as str]
-    [sixsq.nuvla.ui.utils.general :as general-utils]
-    [sixsq.nuvla.ui.utils.time :as time]))
+    [sixsq.nuvla.ui.utils.general :as general-utils]))
 
 (def state-new "NEW")
 (def state-activated "ACTIVATED")
@@ -210,3 +209,10 @@
                                  boolean)]
     (or (str/blank? (:nuvlabox-release form-data))
         (and payload? payload-incomplete?))))
+
+
+(defn form-add-playbook-incomplete?
+  [{:keys [name run enabled type parent] :as _form-data}]
+  (->> [name run enabled type parent]
+       (some str/blank?)
+       boolean))