+++ code/src/cljs/sixsq/nuvla/ui/infrastructures/events.cljs
@@ -6,8 +6,8 @@
     [sixsq.nuvla.ui.cimi-detail.events :as cimi-detail-events]
     [sixsq.nuvla.ui.infrastructures.spec :as spec]
     [sixsq.nuvla.ui.infrastructures.utils :as utils]
-    [sixsq.nuvla.ui.messages.events :as messages-events]
     [sixsq.nuvla.ui.main.spec :as main-spec]
+    [sixsq.nuvla.ui.messages.events :as messages-events]
     [sixsq.nuvla.ui.utils.general :as general-utils]
     [sixsq.nuvla.ui.utils.response :as response]))