+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Added
+
+  - NuvlaBox - new options for opting in on host-level management, at NuvlaBox creation time
+  - NuvlaBox - new tab for managing playbooks
+  - NuvlaBox - new buttons for playbook management operations
+
 ### Changed
 
   - Dependencies updates